Fixed chmod issue when building on Win32.  I completely redid how the distros get packaged. Instead of building a staging directory which then gets packaged, the packaging all happens "in place".  Then the staging directory is expanded out from the final distro, but only for the platform you're currently running on.  This actually builds faster.  The only downside I know of is that I can't share much code between the linux, mac, and windows distro build files. Maybe there's a more elegant solution to this.
I also redid how the samples are built, to make it easier to package up the final distros.
While I was in there I made sample building a lot smarter about not doing rebuilds, thanks to
some of the sweet stuff in ant-contrib.

Finally, fixed ApplicationCreator to merely warn if chmod fails.



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@188 8db76d5a-ed1c-0410-87a9-c151d255dfc7