Fixes issues 242, 540, 933.
Fixing issue 242 allowed for covariant array types to be used in RemoteService method signatures and in the fields of serializable types.  

The logging produced by the RPC generator was significantly improved thereby addressing issue 540.  The root causes for a type's failure to be serializable are clearly identified in the log along with the path which got you to the failure.  Warnings were added for things like missing gwt.typeArgs javadoc annotations for Collections and Maps, final fields which are excluded from serialization, non-static nested types which are not serialized, etc.  

Lastly, fixing issue 933 resulted in a significant decrease in the number of types which actually qualified for serialization.

Patch by: mmendez, scottb (changes to the emulation library)
Review by: scottb



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@1043 8db76d5a-ed1c-0410-87a9-c151d255dfc7