The hidden input used in Safari changes the cursor to a text selector on hover, even though the input element is hidden and behind the main element.  I made reinierz's suggested change to the FocusImplSafari class so the hidden input now has a height and width of 1px, preventing users from placing their cursor over it.
Fixed unreported issue When the cursor moves from one element to another within
the same CustomButton, an onmouseout event is fired in some browsers.  This has
the effect of firing the onClickCancel method, which can temporarily change the
style of the Widget until the onmouseover event is fired.  I added a check to
make sure that the ToElement is not a child of the parent element before
canceling the click. 

Testing: I created a PushButton and ToggleButton and verified that the cursor
on the screen matches the cursor in the CSS style definition.  I also verified
that "internal" mouseout events do not fire the onClickCancel method. 

Issue: 1285
Patch by: jlabanca
Review by: jgw



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@1239 8db76d5a-ed1c-0410-87a9-c151d255dfc7