Fix layout issue where content area was pinned to the right-hand side in hosted mode Linux and Mac/Tiger.
To fix this, we make sure that the styleTesterTimer always runs, regardless of whether or not a stylesheet was swapped out. We need to ensure that the resizing code in onWindowResize is invoked AFTER the stylesheet has been applied, so we add a call to onWindowResize when the styleTesterTimer fires.

Patch by: jamieyap, rdayal (pair prog)
Review by: knorton.

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@2925 8db76d5a-ed1c-0410-87a9-c151d255dfc7