Merging from releases/1.6@3739:3876
svn merge -r3739:HEAD \
  https://google-web-toolkit.googlecode.com/svn/releases/1.6 .

Extensive merge conflicts in JavaToJavaScriptCompiler were resolved by
disabling SOYC -- BobV will re-enable it with his upcoming commit.

A patch for StandardGeneratorContext is also included to avoid a build break,
which will be committed to releases/1.6 at the same time branch-info is
updated, and will be skipped in the revision for the next merge.



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@3877 8db76d5a-ed1c-0410-87a9-c151d255dfc7