Ensure server info from superclasses doesn't stomp on subclass info.
This fixes external issue 3969
(http://code.google.com/p/google-web-toolkit/issues/detail?id=3969).
When a class with superclasses was being serialized, the extra info
for the superclass and the subclass were being stored in the same
place using WeakMapping.  The result was that the superclass version
was seen when deserializing the subclass, causing an exception.  By
using different keys, we can associate each portion of the data with
the appropriate phase of the serialization/deserialization process.

Review by: bobv



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@6983 8db76d5a-ed1c-0410-87a9-c151d255dfc7