Recommiting the new CrossSiteIframeLinker.
This linker works cross-site, because it uses a script tag to
download code instead of XHR. However, like the iframe linker,
it still uses an iframe to hold all the installed code.

This version adds a head element to the inner iframe if one is not already present.

Review at http://gwt-code-reviews.appspot.com/726802

Review by: jgw@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8510 8db76d5a-ed1c-0410-87a9-c151d255dfc7