Delete EditorSupportGenerator, and various tweaks on the way to making addon-gwt recover from the exit of Property from the public API, and the shift to EntityProxyId.
Note in particular that getProxyClass() has moved to EntityProxyId,
which makes it a lot easier to deal with. I think this also paves the
way to a lot of clean up of all the nasty split("@") sprinkled about
the place, by moving it into EntityProxyId, but I don't have it in me
to take it further right now.

Review at http://gwt-code-reviews.appspot.com/912802

Review by: robertvawter@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@8863 8db76d5a-ed1c-0410-87a9-c151d255dfc7