1000: Shortcut for Paste Loses Focus - Desktop Mac
Also fix a bug (probably since early Dec) where adjacent braces in the console do not get highlighted correctly. This was caused by a subtle change in the behavior of DomUtils.splitTextAt, which was caused by a change in the way NodeRelativePosition.toPosition works when the selection offset is exactly between two text ranges.