Add server side support for UI preferences. Restore "Console on Top" and "Plots on Top".
This commit also changes the way FileSerializer persists name/value pairs. The value used to be emitted verbatim, which meant newlines could not be in persisted values. Now the value is encoded using JavaScript-style string literals with escapes. For backcompat, when reading persisted values, if the value is surrounded by double-quotes then JS unescaping is performed; if not, it's assumed to be an old-style value and returned verbatim. This introduces a possible edge case, where an old-style value happens to look exactly like a parseable JS string literal. In this case an extra unescape will be performed.