Do not remove type information before final prune.
If you use -XdisableCastChecking, CastNormalizer REMOVES cast operations from the AST entirely; this causes loss of type information just before final prune, which is really bad.  We should never lose information that way.

Ordinarily, CastNormalizer replaces casts with a call to a cast checking method, and the call site is specially handled to make sure the result type of the method call is the cast type.

http://gwt-code-reviews.appspot.com/1425813/


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10111 8db76d5a-ed1c-0410-87a9-c151d255dfc7