Including the TaskProxy (when available) in TaskEditPlace so we do not do an extra round trip to the server to lookup the task.
Also fixing a bug where the "Task List" menu item isn't selected when the task list is visible because TaskListPlace is no longer a singleton.  We now do an instanceof check instead of an == check to compare the TaskListPlace in the menu item to the TaskListPlace from the place change event.

Review at http://gwt-code-reviews.appspot.com/1428810

Review by: rchandia@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10134 8db76d5a-ed1c-0410-87a9-c151d255dfc7