Adding Column#set/getCellStyleNames() to specify style names to apply to individual cells of a CellTable/DataGrid.  Users can set a general style name to apply to the column, or override getCellStyleNames() to provide style names based on the row/cell value.
I added the methods to Column instead of creating a CellStyles class (similar to RowStyles) because users will most likely want to choose styles based on the Cell value.  If we provide one CellStyles for a CellTable, then it won't be parameterized to the Column type, making it difficult to get the cell value out of the row value.

Review at http://gwt-code-reviews.appspot.com/1446817

Review by: rchandia@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10309 8db76d5a-ed1c-0410-87a9-c151d255dfc7