Fixing IsRenderable API step 2: pass in a stamper instead of the actual ID.
As we discussed a few weeks ago, the goal here is to make the exact scheme by
which elements are marked and retrieved opaque to the IsRenderable
implementation. I thought I went about this in small steps:

1-) Fix the API and UiBinder to use a RenderableStamper class. This class
supports stamping SafeHtml and ElementBuilder, but also exposes the token
explicitly for backwards-compatibility (to keep the patch at a manageable size)
2-) Fix all uses of RenderableStamper.geToken() to use either SafeHtml or
ElementBuilder. In particular I'm worried about performance regressions if we
use SafeHtml.
3-) Profit.

Review at http://gwt-code-reviews.appspot.com/1472801


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10421 8db76d5a-ed1c-0410-87a9-c151d255dfc7