Prevent error storm when NativeWindow.get().focus() fails during ModalDialogBase load
We just ran into a situation where for some reason a DOMWindow didn't have a focus method, causing ModalDialogBase::onLoad's allActiveDialogs_.add(this) not to be called. That results in ModalDialogBase::onPreviewNativeEvent throwing an error on every browser event (such as moving the mouse!), which filled up our error logs.