Changes to keyboard handling in desktop mode
Previously, in desktop mode, many keyboard shortcuts were handled by Qt, by way of the keyboard shortcuts being added to QAction objects which appear on menus. This is problematic because the keyboard shortcut handling is thus not managed using the same code as web mode; most seriously, when modal dialogs are shown, the keyboard shortcuts still work.

Rather than try to keep the QActions in sync with commands, which seems like it would be very chatty, this commit changes it so the QActions generally don't have shortcuts assigned to them--except while their menus are showing.