Fix further rpostback path problems
This problem only happened for me on Windows x64 (username had a space in it, probably not relevant?). Basically in the rpostback scripts, $0 was a Windows-style path, with C:\..., whereas dirname really wants it to be /c/...

Worked around this by replacing dirname with a cd/pwd combination. Seems to work OK.