Fix review pane selection bug
Symptom: When item is preselected in review changes dialog, the first click on it actually unselects rather than keeping the selection stable.

The bug was caused by StatusAndPath not being suitable for a hash key, since it is a JavaScriptObject. It turns out that MultiSelectionModel relies on the objects being put into it being suitable for hash keys, even though it takes a ProvidesKey. For the setSelected(T, boolean) method, the T object is used directly as a hash key.

The fix is to have a non-JSO representation of StatusAndPath, and use that everywhere except when a JSO is truly needed.