Adding an option to change the debug id property and prefix. Some users prefer to use a random attribute instead of the element ID, or they do not want the "gwt-debug-" prefix. I added two static methods, DebugInfo#setDebugIdAttribute() and DebugInfo#setDebugIdPrefix(), to make it easy to configure these settings.
An alternative approach was to create a new implementation of DebugIdImplEnabled, which would not use a prefix. The problem is that switching to that implementation requires the user to add a gwt.xml rule, and its less flexible because you cannot set an arbitrary prefix.  Users who are comfortable with deferred binding can always override DebugIdImpl (as many already do) to control how debug IDs are applied.

Issue: 4179

Review at http://gwt-code-reviews.appspot.com/1618804

Review by: unnurg@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@10806 8db76d5a-ed1c-0410-87a9-c151d255dfc7