Low-level console output now handled by consoleio
consoleio.exe now sends low-level console output to stderr. Since the API
for grabbing low-level console output gives us not streams but snapshots,
each time console output appears the entire contents of the console (up to
the cursor) are sent to stderr, prepended by a \f character.

This makes svnHistory requests with svn+ssh passphrase/password prompting
work correctly except that after answering the prompt (either manually or
by the PasswordManager automatically supplying the answer) the console
progress dialog briefly flashes up. This is because hitting Enter causes
the console output to change, causing consoleio.exe to send a snapshot to
stderr, causing ConsoleProgressDialog to see new output, causing it to
display itself.

Need to figure out a way to detect when a snapshot of a previously answered
prompt is sent. I don't have an answer for this at the moment.