ConsoleIO now streams back most output
The snapshotting-only method of output was causing the console progress dialog to appear on Windows even in cases where the prompt could be satisfied by the password manager. This is because any change in the output (including a trailing \n after passphrase was sent) would cause the entire console output snapshot to be sent, which would trigger the dialog.

I modified ConsoleIOMain so that in the (very) common case that each new snapshot is just a longer version of the previous snapshot, only the difference is sent. This should work fine except in the case of scrolling or if \r et al are used. In those cases ConsoleIO will revert to snapshotting. Note that this will be very inefficient if the screen is being scrolled (though no more so than before this check-in) since any output would cause the entire buffer to be re-sent.