2199: regression - matching parentheses not deleted
Our monkeypatch of Ace's removeLeft() function broke because removeLeft() was refactored as part of the introduction of editor behaviors. To fix, I refactored our removeLeft() monkeypatch into a behavior (which then has to be monkeypatched into TextMode, unfortunately).

The insert() monkeypatch (that actually puts in matching braces) I left alone since it seems to be working right now, but there's every chance it will also break in the future and then we'd need to refactor that as well.