Support viewing Git diffs using the project's default encoding
Project encoded files need to be translated to UTF-8 for them to appear properly in the UI layer. However, when sending patches back to the server, we need to make sure to translate back to the project encoding before applying, or else you will end up with mixed encodings in the file.

If the diff cannot be translated from project encoding to UTF-8 (probably because the file is not actually encoded using the project encoding) then we effectively just treat it as UTF-8. In this case we remember not to do UTF-8=>project encoding when we apply patches based on this diff.