Do not used proportional fonts for editor
Some desktop users on Linux are experiencing proportional fonts being used as the editor font. This commit fixes the problem by programmatically detecting whether a font is fixed width.

Since auto-detecting the font is potentially expensive (haven't measured but font metrics need to be loaded now) we now cache the results of auto-detection in a static.