Prevent/recover from tabsets that only contain VCS
Includes fix to 2392: WorkbenchTabPanel.selectTab(int) should prevent use of -1

The problem was reported by a user who put the VCS tab on a tabset by itself. This caused the tabset to come up blank in the global environment, which then caused the tabset's client state to say tab -1 was selected, which then caused an exception on every startup thereafter.