Improve autoindentation with continuation lines
See autoindent_test.html, there are a couple of tests that no longer pass.

However, the new test that passes seems way more important.

It is not obvious to me whether the old tests can easily be fixed, and it also seems like those cases are uncommon and the new behavior is not horrible.