2798: Mac - left curly bracket "{" not working
On Mac, the keyboard combo for { on Swedish keyboards is Option-Shift-8. For some reason, Qt sends this key event along in such a way that Ace sees it at Option-Shift-0 which is the keyboard shortcut for "unfoldall", and squelches the keypress. Since this keyboard shortcut is not documented anyway, just get rid of it and fix {.