Adding the GWT 2.4_2.5 API Checker configuration file.  Many API changes were added to the 2.3_2.4 file because we never created the 2.4_2.5 config file, so there are a bunch of API change exceptions that had to be copied.  Pretty normal, but we should get into the habit of creating a new API checker config file immediately after releasing a version of GWT.
I noticed that we have a big API change to the IsRenderable interface between 2.4 and 2.5, which will cause API breaks for anyone using it.  I doubt it is used very much (I think its used by UiBinder to render nested elements), so it should be okay to change it with minimal impact to users.  But we might want to note it in release notes.

Review at http://gwt-code-reviews.appspot.com/1721803

Review by: rdayal@google.com

git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@11000 8db76d5a-ed1c-0410-87a9-c151d255dfc7