DefaultProxyStore violated ProxyStore#nextId contract. Contributed by tbroyer.
"The javadoc for nextId states that the return values must be unique, but
DefaultProxyStore returned the same value twice in a row unless an object is
added to the store. This situation happens when serializing a list of
ValueProxy objects (containing at least two distinct objects), which
unfortunately wasn't tested."

Fixes issue 6961.
Review at: http://gwt-code-reviews.appspot.com/1622803/


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@11006 8db76d5a-ed1c-0410-87a9-c151d255dfc7