Fix for Issue tracker issue #5739
CfgBuilder does not add a graph edge from a cast operation to the method exit or
nearest catch block, so casts always succeed as far as the flow graph is
concerned.

This patch adds a visitor for JCastOperation to add a RuntimeException exit edge
in much the same manner as the JMethodCall visitor. In fact, a JCastOperation is
an implicit call to Cast.canCast() except for the fact that the replacement of
JCastOperation to a method call doesn't happen until much later when
CastNormalizer is run.

Review: http://gwt-code-reviews.appspot.com/1849803/


git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@11332 8db76d5a-ed1c-0410-87a9-c151d255dfc7