Removes the static book keeping in the Timer class.
This changes removes code that keeps track of Timer statically.
Based on history the code was always there from the beginning but
my guess is, this is no longer needed: New-ish Scheduler API is not
doing the same workaround and also there are other places that are
using Impl.setTimeout directly without the workaround.

Change-Id: Idf6a9c81a5654bc2eb736a2b4ef81d5cf0517728