Update how columns are hidden in CellTable.
Hide columns that shouldn't be visible by setting display to none.
This is to fix a rendering issue in safari and chrome, where the column
is hidden, but doesn't allow the other columns to expand to fill the
remaining space.

Original patch by: Aliseya Wright

Change-Id: I49fd531a02ca0a1a55eabba5548670ad86273a43
Review-Link: https://gwt-review.googlesource.com/#/c/3130/