Do not initialize reference fields with null.
Leave all uninitialized fields with an undefined value instead of initializing them
with null.

In the past in some cases fields where initialized with null and some where left undefined;
commit 202e5a5 made all scalar fields to be initialized increasing codesize by 0.25% (and most
of the increase was in the initial and leftover fragment if runAsync was used). Non primitive array
elements on the other hand, are always left undefined.

This patch uniformly leaves all uninitialized fields as undefined.

Change-Id: I423852b1171a15c0bd187aa0eec4f527b3f95643