Revert "Add show relative to Element in PopupPanel"
This reverts commit 7b6d727fa6b53eaf42aa17d499582a17d967b071.

The added method is marked as final and it is quite common that classes which are extending PopupPanel to have already introduced the method. So the breaking impact of this patch is big and we didn't anticipate it before.
We should discuss it again if/how we want to introduce it.

Change-Id: I373a364510e0c8b477e5559727db392f3cbdf674