Add "superdevmode" rule to Ant files generated by WebAppCreator
As a result, the sample apps that use Ant will have a "superdevmode"
rule.

Verified that it works with Hello. It currently doesn't work with
Showcase because ShowcaseGenerator assumes that it can read
Java source code using the classloader; we'll need a different API
for reading source from the GeneratorContext.

Change-Id: I059bf8eaffd0b28ebd0c34088edbed690b44b5a6