Fixes namer output for package-info classes.
JDT compilation of package-info.java classes results in the
creation of AST name nodes containing "package-info".
"-" is an illegal character in function names in output
javascript. So the javascript namers have been updated to
fix it.

Previously this was not a problem because this "-" character
was only in package-info classes and the Pruner was
removing those classes. But separate compilation does not
run the pruner, thus revealing this issue.

Change-Id: Ia322dadd36f2e117a3024cf3ed0f2cee8bfcb873