Set SO_REUSEADDR when starting Jetty
SO_REUSEADDR allows us to listen on TCP ports that are still in
TIME_WAIT status (i.e., someone else recently finished using it),
which is necessary for busy testing environments where many (or even
all) ephemeral TCP ports may be in this state.

The original comment says SO_REUSEADDR was disabled to prevent sharing
a port with another process.  However, it's actually SO_REUSEPORT that
allows multiple processes to listen on the same port.

Change-Id: Iafce162ef0398b7c935491a37a1b02cb0e48dfd8