Added missing fileset mark to HTMLTable.gwt.xml
All the modules in the standard library are aggregated together
in this way during separate compilation. The marking was
missing from HTMLTable.gwt.xml because the file was added in
a different review.

Since HTMLTable.gwt.xml does not properly declare all its
dependencies it can't be compiled on its own, so this marking
is necessary for correct separate compilation of the User library.

Change-Id: I1dbf7aa538fb819a7413a5efeb7616773137fb43