Fix bug: Object sizes > 2GB show as "16 TB" in environment pane
The cause of this bug was subtle--in GWT, the "long" type is actually a (pair of
ints)[http://www.gwtproject.org/doc/latest/DevGuideCodingBasicsCompatibility.html],
but it doesn't work quite the same way as you'd expect and consequently
doesn't play well with native ints returned from JSNI methods. Any
coercion of a JSNI int to another type of value generally results in
truncation to Int32.

The fix is to remove the use of the GWT-implemented long in most places
that deal with large object sizes and to use "int" instead, which GWT
largely leaves untreated and unbounded (except by the underlying
JavaScript's size limits, which are plenty big even for large file
sizes). The downside to this approach is that care must be taken to
avoid GWT forcefully truncating the values.