fix window cycling via shortcut key on Mac desktop
This reverts a fix originally intended to prevent a bug on the Mac
desktop wherein unrecognized shortcuts weren't correctly ignored. Now
that the Mac desktop app is wrapped in Cocoa, we don't need this
workaround; it has the side effect of causing us to eat events
associated with system shortcut keys modified with Cmd.