fix bug: negative scrollbar height during boot (superdevmode only)
Some GWT internals wind up with a negative height while calculating the
height of a scrollbar; the problem is timing-dependent and doesn't
happen while stepping over the calculation in the debugger.

Because the errant calculation is several calls into GWT and difficult to
influence from RStudio code, this "fix" sidesteps the issue by deferring
the creation of the offending DataGrid until the event loop is finished.
The fix also has the side effect of avoiding the unnecessary switch
between grid and list view that formerly happened on boot when the user
pref was set to grid view.