Computes JS strong name earlier, to avoid recomputation.
When linking the strong name of the output JS is needed to uniquely
identify the output file.

Calculating this hash of a potentially large String is expensive. When
compiling monolithically this hash will only be calculated once, so it
is mostly OK that in the past this hash was calculated at the last
moment.

But when compiling (and linking) separately, the output String is
larger and made up mostly of chunks that have been previously seen and
are being reused. Taking the hash of the large String that is the
result of combining these chunks, starts to be extremely wasteful. By
having precomputed hashes for each chunk, it becomes much cheaper to
generate a unique hash for the set of chunks (by hashing just the
combination of precomputed hashes). This change prepares the way for
this eventual separate compilation related change.

Change-Id: Ied09ca9f3e73143babc890e49d27456de7054ada