teaches CompilationResult to combine library PermutationResults
In the past the StandardCompilationResult artifact acted as a facade
for a single PermutationResult instance as it surfaced JavaScript code
and StatementRanges to the linker pipeline.

With the addition of separate compilation and linking there are now
sometimes multiple PermutationResult instances (one root "application"
instance and multiple "library" instances) that make up an application.

This change adds an optional constructor to StandardCompilationResult
to pass it an extra ordered set of "library" PermutationResults and
updates JavaScript code and StatementRanges retrieval (and some other
things) to return the combination of all of that data from all of the
known PermutationResult instances.

The net result is that all of the JS of an application and a set of
precompiled libraries is combined together in the final linked output.

Change-Id: I68728bd94b9daaf805ef48ad6b6b7f2903b485aa
Review-Link: https://gwt-review.googlesource.com/#/c/5923/