Put JavaScript globals into namespaces (take 2)
(There is no user-visible change since it's not turned on in this patch.)

This is to work around a Chrome limit on the number of global variables
when attempting to run a large GWT app in Super Dev Mode. It should also
make the debugger somewhat easier to use because not as many variables
show up under "window".

Implementation: a new compiler pass creates a separate JavaScript
namespace for each Java package. Each namespace contains the functions
and global variables for that package.

By default, package namespaces are generated by taking the first initial
from each name in the package path, so for example, "com.google.gwt"
becomes "cgg" unless that name is already taken.

TODO
 - enable for Super Dev Mode
 - add a flag to the regular compiler.
 - Tweak the pretty namer so that it doesn't rename globals as often.
A variable that's moved to a namespace only conflicts with other
variables in the same namespace.

Bug: issue 8532
Change-Id: I9f20d247ecf932dbc6c805ea43648d81241b2ef2