Fix script-injection tests, now that xsiframe is the default linker
Resets the linker to 'std' for those tests that use a <script> in the
gwt.xml module descriptor file. Ideally, we'd probably want to be able
to use a specific host page, or configure it for the test with additional
script tags.

Also, client coverage seems to only work with the std linker.

Change-Id: Ibea7181792a68af9cc3a0fa2d887098297eeb0fe