Synthesizes a more compatible EntryMethodHolder.init().
Old EntryMethodHolder.init() synthetisation repeated GWT.create()
handling and verification that was already present in UnifyAst. This was
not only wasteful but was incompatible with separate compilation. This
replacement version synthesizes just GWT.create() calls and lets
UnifyAst handle the interpretation and verification of the calls.

Switches names of classes that are generated per-module, module names
stored in libraries, and name references to other libraries to be based
on the module's canonical name (to avoid collisions resulting from
multiple modules with the same rename).

Fleshes out the LibraryJavaToJavaScriptCompiler.ensureFullTypeLoaded()
call to match the correct type loading pattern of UnifyAst.findType(),
and comments on the desire to remove the duplication of logic.

Change-Id: I6a3150a248903e7d3b6ddec421248596aa57c1a4
Review-Link: https://gwt-review.googlesource.com/#/c/6745/