Super Dev Mode: faster sourcemap loading
Instead of parsing the sourcemap as json and writing it out again,
replace a template variable while streaming the file.

To support this, I gave the GWT compiler the ability to set the
base URL in the sourcemap. (There's an option for this but it's
not exposed as a flag yet.)

Also, cleaned up the code server's internals and sourcemap
generation. (We no longer need to subclass SourceMapRecorder
to populate the "names" field in the sourcemap.)

Change-Id: I51ff3f0970aadb832b8f430175b103b104e8fd62