Replaces 'in' checks with 'gets' in InternalJsStringMap.
Andrei suggested in an earlier patch to use 'undefined' for avoiding
double look-up (contains + get). This requires ensuring that we always
convert 'undefined' to null while putting values so we thought it
would not worth the effort. However later in my tests I noticed that contains
implemented with 'in' is much slower than accessing and comparing the value
to 'undefined' both in Chrome and Firefox. So there is much more potential
than just saving an additional lookup (i.e. up to 3x speed!):
http://jsperf.com/object-keys-iteration/47

This patch basically gets rid of 'in' statements in InternalJsStringMap
and also simplifies InternalJsStringMap a bit further by sharing more
code between the modern and the legacy implementations.

Special thanks to Andrei for the idea and also debugging code together
with me in the GWT meetup :)

Change-Id: I6b3a02916931c89901635c2e35c1b972b15531b5