Adds invalidation of referencers of types that change JSO status
When type Foo calls methods on interface Bar the signature of the method
dispatch is a reflection of the JSO status of interface Bar. The JSO
status of interface Bar can change even when the Bar file on disk has
not been modified (as a result of the creation or deletion of other
types that implement Bar). Thus the addition of this staleness
detection.

Change-Id: I79b74c3ed52e24f576964c031899b8d65c972ad0
Review-Link: https://gwt-review.googlesource.com/#/c/8661/