SuperDevMode: try to reuse the first compile if nothing changed
To do this we need to make sure we get a cache hit on the
binding properties in the InputSummary. Removed an unnecessary
binding property (it's set later) and stored the actual properties
chosen (instead of the binding properties requested) to make a
cache hit more likely.

If the user isn't using Chrome + English or there's another binding
property, they won't get a cache hit. A workaround is to use
-noprecompile. (We might also want to add a way to configure
the binding properties for the first compile, but that's for another
patch.)

Change-Id: Ifdbe5515ca29a3fc3e5e3990d922948ff04a10ba