Revert "Stops reparsing the ModuleDef on every refresh in SDM."
The problem is that Recompiler.maybeSetBinding() is not
reversible, so on the next compile we can't set the binding
to a different value. The recompiler is using setAllowedValues()
to force one permutation and this needs to be rolled back
somehow before doing the next compile.

I wonder if there are any other changes that might leak
into the next compile? It might be safest to clone ModuleDef
before modifying it.

This reverts commit f743cb1815e68309e1d2e97c8817d4dc89da67d4.

Change-Id: If9862defa8f119a6e007e0fb70738bd520e3f37e