Marks types generated by invalidated generators as "modified".
Previously in per-file compilation mode when a resource was seen to be
modified the generators that read this resource were marked stale and
care was taken to make sure they were rerun. This was adequate to ensure
that any new types (that weren't created on a previous run) would be
fully processed. But it was not adequate to cause types that had been
previously generated (but which might now have modified contents) to be
reprocessed. Since only types believed to have been modified are
reprocessed, these types are now added to the modified list.

Change-Id: I7f3fe0aa6239e43db1c7f1453d662cc04ef46c2a