Fixes UnifyAst.computeOverrides() when loading a partial world.
Normally compiles load all referenceable types, and so pruning non-live
methods does not change the behavior of computeOverrides().

But when loading only some subset of types it's possible for a method in
a parent class to get pruned (because it is not referenced in this
limited world view) which was making overriding methods in the subclass
appear to not be overrides.

Fixed by moving computeOverrides() to before
pruneDeadFieldsAndMethods().

Change-Id: Ia7248de717f5fc51b5feb1c64783cf4e28d9a140