Fixes for JDK8.
Some changes in Java 8 make tests fail. In particular there
are small changes in regular expressions that RegExp and
String.split, and also changes to the collection contract
when dealing with null keys.

This patch fixes the tests so that they run under Java8 and
two issues were opened to update the emulation library to
implement Java8 semantics (issue 8912, issue 8913).

There is still a pending failure in tck, issue 8914.

Change-Id: I2ddf90b84ec0e1e583983e1a6b7057c83d42773a