Refactors LinkedList to not use circular list.
The new implementation is using double pivot instead of
single.

This is to make LinkedList safer when preconditions are
not enabled and iterator is not used properly.
For example, with a circular list, if you keep calling
next/remove you might never hit an NPE because it will
eventually come back to beginning and start iterating
again.

This might be a little bit on the paranoid side but given
the new implementation is no more complex, it is a good
trade-off.

Change-Id: I4915d5c7b3a6a0e48b96c14be4ce0d64bf3b0de2
Review-Link: https://gwt-review.googlesource.com/#/c/9363/