Allows linker to include comments in JS code to be prefixed.
Linkers might add JS code at the beginning of the compiler output and
it is transformed into a one liner so that the (already generated)
sourcemap does not need updates.

These prefixes were assumed to be simple programs with no trailing
comments (i.e. single line comments) that prevent straight
concatenation of the JS code lines.

This patch removes the restriction and strips these comments assuming
that the JS program is are well formed.

Change-Id: Ib6945e9c4d46cad9287475b7336a32e23da608ac