Fixes a bug in JsStaticEval.
JsStatic eval used to change the order in which binary operations
are performed in an attempt to reduce the number of parenthesis
required to represent expressions.

The method to perform this transformations is buggy and poorly
written. This patch removes this method (correctness is
much more important than codesize). In particular completely
ignores the associativity of some of the operators it is moving
around. Also arithmetic operations should not be reordered as in
floating point arithmetic reordering might affect the result.

This patch simplifies reordering and only considers real associative
operations like COMMA, logical and bitwise AND and ORD. The
reordering does not consider arithmetic operations nor string
concatention anymore.

Bug: issue 8934.
Change-Id: I66f0a7c770ac376f530003acaf171e3a99815160