fix indentation following a ','; align =
This commit fixes indentation following a ',', and also provides a
feature whereby multiple assignments following a ',' can be
automatically outdented and aligned.

For the first case, we now indent following a comma when appropriate,
e.g.

    int x = 1,

        ^

and in the second, we automatically align '=', e.g.

    int x = 1,
      y =
        ^~|

by adjusting the line's indent such that the '=' align.