Changes OutputFileSet to overwrite on equal timestamps.
Previously OutputFileSetOnDirectory took special care to not overwrite 
a file in the output directory if it had the same timestamp as the 
current artifact being written. For a long time this had no effect
since Link.java deleted al output directory contents anyway and DevMode
was  not using OutputFileSets to perform it's writes.

But somewhat recently DevMode was changed to use 
OutputFileSetOnDirectory and it resulted in a subtle change in behavior
where generated files from 1 compile would not be overwritten by the 
samely named (but differently contented) generated file on a subsequent
compile.

As a result running a prod mode test followed by a dev mode test would 
fail as the dev mode test tried to reuse the incompatible prod mode
test Foo.nocache.js file.

Change-Id: I0916ab7933f9dfcb36b361f2e21657fac8174d67
Review-Link: https://gwt-review.googlesource.com/#/c/10311/