Make call sites and method agree in number of parameters after pruning.
When pruning the last few parameters of a method, the pruner would
allow for extra arguments in calls, relying on JS semantics for
calls in which calls might pass more parameters than are declared in
the method declaration (which will on most cases be ignored).

Having a mismatch in the number of arguments implies those calls won't
be inlined (by either fo the inliners).

Change-Id: Ie302158e35ebef3ded0f037eccd1c3c4f6f63ee4