Restore gwt.persistentunitcachedir for Compiler, DevMode and JUnitShell
Change Ib995c18fe6842dc8914457ceff43e17808bfa550 changed the behavior
of UnitCacheSingleton.get() to only use the gwt.persistentuniecachedir
system property when the passed in argument is null, whereas the
previous behavior was to use the argument as a fallback with the system
property having precedence.
This change restores support for the system property in selected entry
point classes (i.e. not SDM, neither LibraryCompiler) by using the
three-arg overload of UnitCacheSingleton.get() to explicitly pass the
value as a fallback in those cases.

Change-Id: I4ea07f9c0e66bc1636a5b4c6035918c73367633f