fix packrat initialization order
This one is subtle -- the 'deferred initialization' in
SessionPackrat.cpp needs to occur _after_ Packrat's init script has run,
as this does some work of e.g. setting environment variables and so on.
If that isn't set, then the deferred init believes that Packrat mode is
not on, and doesn't hook up any of the file monitors and so on.