Fixes AssertionError in ImplementClassLiteralAsFields.
Jsni methods, in order to expose some internals to Java AST vistiors,
maintain some redundancy for class literals, methods and fields
references.

The code handling class literals had a wrong assumption that
a particular class literal would not appear more than once in the
representation for the the Java AST visitor, whereas it would
actually appear as many times as it does in the JSNI source.

Bug: issue 9043.
Change-Id: I6ecc46197f4bbb2220870ed23c7491bc788a942b