Enable Java8 for GWT on JDK7 with Java8 VM
This patch stubs missing LambdaMetaFactory which is missing
in JDK7 but referenced from Java8 byte code for lambdas.

With this patch one can run java8 on jdk7 as long as it is
running on a Java8 VM. Java8 VM still required because
generators might try to access compiled code with reflection
which would not work in Java7 VM due to incompatible class
format.

Change-Id: Ie1abc7aee32f9940292b750863d0639b5c386ff9