redraw data viewer lazily (fix refreshing structure on Qt platforms)
On Qt platforms, it's occasionally the case that changing the structure
of a data frame causes it to blank out in the viewer. This is a result
of us doing a full refresh when the structure changes, and some Qt
networking issues that occur when the frame reloads its resources (which
we haven't fully resolved).

The fix is twofold, and should improve performance of the data viewer on
all platforms:

- Formerly we aggressively reloaded the data viewer whenever data or
  structure changed. We now do this lazily--that is, the viewer doesn't
  refresh anything unless it's the active tab, and refreshes are queued
  until the tab becomes active (if that ever happens).

- It's now possible for the viewer to refresh the structure of a
  frame without performing a full window reload (with all the HTTP
  roundtrips that entails); instead, only the affected DOM elements are
  removed and recreated.