Fix a bug in GwtAstBuilder.
The ReferenceMapper object in GwtAstBuilder, which caches name to type
instances, had the wrong lifetime.

That may cause compilation units to be serialized sometimes with non
stub to types outside itself. That resulted in an unified AST with
multiple type instances representing a single type.
Then queries like isInstantiated that are keyed by the type instance
failed and produced incorrect optimizations.

Change-Id: Ie154e7eab257d9c0705a30c2196caec5993657a2