Improves clinit step debugging in SDM.
Standard clinit function behavior is to immediately replace itself with 
an empty function so that only the first invocation does anything. But 
even the empty function involves a couple of unnecessary line steps 
during debugging.

This patch switches the standard empty function for the Object 
constructor function when compiling incrementally. Since the Object 
constructor function is native the debugger will not enter it, which 
gives a better experience.

Change-Id: I0cc66d823281711e989898fc412c26a328a8a016
Review-Link: https://gwt-review.googlesource.com/#/c/11600/