Adds UmbrellaException causes as suprressed
In UmbrellaException, the first exception was used as cause
and the rest was kept as a property. This was causing the rest
of the causes to be not logged with logging systems or required
special handling.
By adding them to surpressed will make all causes logged by default
without any special handling for UmbrellaException.

Change-Id: Ifed95a990b37028930b7f6350ee3f79e81aca224