Fix a bug in method reference in Java8.
When using method references to methods that have the
same signature but from different classes, wrong method was called.

For example, f(A::id); f(B::id); A and B both have method id()
with the same signature, B::id() is not called correctly.

The reason is that the generated names of inner lambda class were
the same since they have the same method signature, thus only one
lambda class was created for A::id and B::id.

Fix the issue by adding class name to the generated inner class
name to identify different methods.

Bug: issue 9040
Change-Id: Ibc6b12d26d473ab32b821bb2202e78ae92670774