Fixes error handling for Resource gathering.
As part of recent file watching change, we introduced a
regression that was causing some resources to be not
resolved.
The main underlying issue is, if there are many file
watchers available, then the file system throws an exception
and when exception occurred we were simply returning an empty
list (instead of doing a full scan as previous code did).
As some part of the Compiler is passing TreeLogger.NULL we
were not observing this issue in the output.

Patch makes some changes to improve the situation:
 - When we try creating a watcher, if it fails, we track that
   and fallback to full-scan. This not only fixes the problem
   but prevents unnecessary retries to create the watcher.
 - Switched System.err logging so that any problems related to
   traversing of resources are always communicated to the user.
 - Adds System.gc to CompilerTest#setup so that we do some
   cleanup of ResourceAccumulators between test runs to
   reduce the number of file watchers.

Change-Id: I7ca25a10429747c0b4ac5c41af247530d3d91c5d