Fix bug for method references used as lambdas.
An assertion was thrown when the same method reference was
used as a lambda expression in different compilation units.

The compile was generating the an anonymous class with the
exact same name in different compilation units which
resulted in two copies of the same class in the unified AST.

This patch fixes the problem constructing a unique name by
prepending the compilation unit main type to the anonymous
class name.

Also in this patch a some needed refactorings.

Bug: issue 9190.
Change-Id: Ia20051a28406772070f9b7b231dcb0d45ca296e9