Fixes Cookies.getCookie() to prefer the most specific cookie instead of the least specific cookie.
The cookies that come from the browser are sorted by path order so that cookies
with the most specific path appear first. If multiple cookies with the same
name are present we should return the most specific cookie. As it stands, we
end up returning the last cookie which is the least specific one. This changes
it so that we return the first cookie, which is the most specific.

Change-Id: Ie8c76c349780f8c59beb98c4df2a3497e78e7216