eagerly tokenize when moving to position
This should fix errors when building the scope tree (as TokenIterators
are used when navigating the tokenized document).

Previously, it was possible that we would request the scope tree while
the background tokenizer was running (thereby encountering the document
in a half-tokenized state); the fix is to eagerly tokenize up to the
requested row so the TokenIterator never attempts to traverse a
non-tokenized document.