Make @JsFunction work with browsers that don't support '__proto__'.
Note that the implementation is not equivalent to setting the prototype.
For example, 'instanceof' will not work and if new properties are added
to the object, they will not be reflected in the lambda.
I think both of those shouldn't be a big problem though.

Change-Id: I2324b558d276210bdd394466f141f6fc3196dc40
Review-Link: https://gwt-review.googlesource.com/#/c/13390/