Remove JsInliner limitation to keep field accesses.
JsInliner avoids removal of unused expression that are purely
de-references. Our guess is, it was working this way to preserve
any potential NPE on dereferences. However java optimizations
already not trying to preserve such NPEs so we don't find any
reason to have it in js optimizations.

Change-Id: I055a83019fff17b4358a107e01eaa1d1bb0a193e
Review-Link: https://gwt-review.googlesource.com/#/c/13521/