Allow safe inlining in the main scope and fix a subtle bug.
JsInliner would not hoist a variable if it only appeared in a
var definition, i.e.

  f() { var a = g(); }
  caller() { f(); };

would was inlined as

  caller() { a = g(); }

Also, this patch removes a hack that allowed inlining only in the scope
of certain function calls. It nows allows inlining in the global scope
as long as no variables are extruded.

Change-Id: I44fd3101bd3c9b420bb12b0a2630ee8068e08054