re-work bg tokenizer call
This fixes an issue where the background tokenizer was a bit _too_ lazy
-- for example, when attempting to create a C block comment, e.g.

    /**
     * foo
     */

It was possible that inserting a newline within the comment would
not be properly eagerly tokenized, and so the resulting line + state
was not inferred as a comment (and this caused problems with e.g. auto
insertion of continuation stars, e.g. ' * '.)