Implement JsOpaque check.
The purpose is to prevent from passing unmatched variables
at JS side accidentally.

For example,
@JsType
interface MyJsInterface {
  void foo(A a);
}

If A doesn't have any accessible methods from javascript (
not one of the following types: Js-doced Type, JSO, JsFunction,
non-long primitive types, Boolean, Double, String, Object),
but can be referenced externally by JavaScript (exported field,
parameter and return type of exported method), it is a warning.

But if this is intentional as the parameter will be only treated
as an opaque object (e.g. a Map), it should be annotated with
@SuppressWarnings("JsOpaque").

@SuppressWarnings("JsOpaque") can be annotated on a method, which
avoids JsOpaque checking on parameter types and return type, on
parameters and on fields.

Change-Id: Ic2fe9efb520d6350716add31d0172f455c524791