Rollback field pruning change for new flags.
Recently JField.canBeReferencedExternally updated so that
it doesn't rely on exporting flag. Actually that wasn't the
plan with exporting flag. We want to keep everything pruned
if the flag is not set. By this way, it will prevent common
code from effecting apps who doesn't need exporting.
And without exporting native JsTypes and JsFunction still work
which covers the use case of a lot of apps just looking for
simple interation with JavaScript.

Laters exporting flag will be enhanced to take parameters for
fine tuning what is exported.

Change-Id: Icea66c08173ccf1f08ef2d7811085abf757c6fcc