provide top-level file completions in unmonitored dir
Previously, we relied on the file monitor always for providing file
completions when within a monitored project. However, certain files /
folders will be excluded from the file monitor; we still want to produce
(top-level, non-recursive) completions within that directory, so we make
sure that we take the union of what the file monitor gives us, alongside
a simple 'list.files()' listing.