Make NativeSubclass.class.getSuperclass() return Object.class.
There is an assumption on the ordering of the output that code for
a super class is emitted before that of the subclass in the JS
output. Because of the replacement of Native class literals for
JavaScriptObject class literals, the invariant was broken and
in some cases the produced output was errorneous.

Change-Id: Id0f5b8a536379d7307a59cac8a4bd2c8fab34d94