Make each method reference be its own different lambda.
Even method references to same method (with or without explicit
qualifier) might need different impelementations due to generics.

Instead of attempting to reuse, this patch takes the simplest approach,
which is to consider each method reference a different implementation
in the same way lambdas are treated.

A future optimization could be introduced to deduplicate lambdas
and that would not only optimize member references.

Bug: #9333
Bug-Link: http://github.com/gwtproject/gwt/issues/9333
Change-Id: Ifef2bb703ad4172977e965f786a0807bdccd5f7b