Add support for custom help handlers
This PR defines a mechanism by which implementors of the .DollarNames method can provide custom help content for completions and custom urls for F1-based help requests.

To provide custom help content for completion a "helpHandler" attribute is added to the results returned from .DollarNames. The helpHandler is a function which can produce custom help info and URLs for a given symbol.

To provide a custom URL for an F1-based help request an S3-style method ("help_url_handler") is provided for a given class and F1 help requests for symbols of that class accessed via $ are routed to the handler.