adjust scroll position to compensate for line widget height
This change fixes a bug wherein the scroll position is set incorrectly
when a document contains line widgets, affecting navigation actions in
e.g. popin/popout, and jumping to lines via Find in Files.

The essential problem is that, when a line widget is added or grows,
content beneath it is pushed down. If the line widget happens to be
above the cursor, the cursor is pushed down with the rest of the
content, so it's easy for the cursor to be pushed out of the viewport
entirely.

There are three components to the fix:

1. Most importantly, manually adjust the scroll position in order to
   hold the position of the cursor constant.

2. When a document contains line widgets, don't attempt to navigate
   using a raw scroll position, as the scroll position includes the
   height of rendered line widgets, which will only be valid at an
   indeterminate time when all widgets are rendered.

3. Defer the restoration of the scroll position, as the computations
   needed for (1) cannot happen reliably if the scroll position is
   manipulated prior to the initial render. (Most navigation already
   defers the restoration of position.)