Fix bug where reopening a shutdown session didn’t start new shell processes for existing terminals.
When a session was terminated (e.g. File->Quit Session), the flag that tracks if the terminal session has a running process was not being set to "false". On reopening that session and reconnecting that terminal, the server thought there was already a running process and did not start a new one, leaving a dead terminal.

Fixed by splitting serializeConsoleProcs() into two functions, rather than one function with an optional argument. New function is markConsoleProcsSuspended(), and is now being invoked during suspend AND during shutdown handling.