remove custom pointer cursor for linux desktop
On some older versions of Ubuntu the default hand cursor was quite ugly, so we substituted a custom hand cursor. A couple negatives of this were that we can to constantly remember to use the custom cursor and there were always parts of the product (e.g. the embedded help viewer) that used the system cursor.

Ubuntu 16.04 appears to use a better looking hand cursor, so this change may no longer be required. In any case, I think we may have over-stepped here in the first place since the user is living with the "ugly" cursor in every other app including the web browser so it's probably either not bugging them or just white noise.