Only write QuitChildProcessesOnExit when it's not "Default"
This is consistent with the behavior of other project options we've introduced outside of the core ones.

The rationale is that if allow new values to be written then user's .Rproj files are constantly getting diffs when they upgrade versions. This might be okay in some cases, but for options like this one which are advanced and disabled by default it will likely just stir up questions / support incidents.