Filter NA attributes from driver enumeration in the odbc connection pane
unixODBC has a bug that SQLDrivers does not properly return driver
information if it is specified in the user configuration file
`~/.odbcinst` rather than the system wide configuration file. This
behavior was reported upstream and should be fixed in the next unixODBC
release (2.3.5).

As a result of this bug drivers specified in `~/.odbcinst` will have
`NA` values when returned by `odbc::odbcListDrivers()`, causing a number
of Warnings and empty entries added to the connection Dialog. See
<https://www.dropbox.com/s/wx64d9e6frtvbk3/snippetsFile.gif?dl=0> for an
example of the current behavior.

To fix this we can simply filter out NA entries from the dialog,
avoiding the issue.