Switch to alt-buffer after refresh with full-screen program
If a full-screen program is running in a terminal, and the browser is refreshed, that terminal would then be showing the full-screen program in the regular (scrolling) buffer.

Fixed by detecting this case and sending the sequence to switch to alt-buffer after reloading the regular buffer.

Some cleanup to improve consistency of metadata handling in TerminalSession class.

Rename XTermWidget.altBufferActive to xtermAltBufferActive to differentiate from the server-side flag.