Set zombie flag directly on server, not via RPC
The zombie flag is set when a process exits on server, based on the ConsoleProcs settings and the user preference. Was doing this from the client in the onExit handler, then sending the server an RPC to set the bit, but this is unnecessary as the server already has the info to set this directly.