Open source commit for https://github.com/rstudio/rstudio-pro/pull/155.
Fixed MacOS socket timeout bugs. This was being caused by setting the global variable R_wait_usec. Upon investigation of the R code and through some debugging, I discovered that this setting does not appear to be needed. R only really uses this setting to control IO wait timeouts, and is not applicable in our use case.

I believe the intent of this setting may have been to keep the session responsive while R is executing code, but this is not the case. I verified from debugging and code inspection that regardless of this setting, our PolledEventHandler is called consistently when the session is inactive, and is not called while executing R statements. However, R contains code to allow processing of the PolledEventHandler during some long processing, such as within loop statements that again gets called regardless of the R_wait_usec variable.