deny directory listings in GWT file handler (fixes #1561)
Prior to this change, going to a URL which referred to a directory (such
as server:8787/js) resulted in HTTP/500. This is because the file
handler attempted to read and serve the directory as though it were a
file. The fix is to check for directories up front, and serve a 403 if
we're asked to serve one (by convention for Web servers with directory
listings disabled).