pass all arguments to the rsconnect package by name
We've had at least a couple of instances where the contract between the
IDE and the rsconnect package was broken unintentionally. The cause is
systemic:

- The contract is implicit; the rsconnect package does not know which
  functions the IDE calls, so it is not obvious when changing the package
  whether a change to a function signature might break the IDE.

- The IDE calls all functions with positional arguments, making it
  impossible to change argument orders or insert new arguments anywhere
  but at the end.

It'd be nice if we had a formal way to test the contract, but there's
certainly no reason we can't start calling named arguments rather than
positional arguments to help making the implicit contract less fragile,
so that's what we're doing here.