rename data viewer resources for cache busting; fixes #1573
Some RStudio Server users have reported that they see a mix of white and
dark in the data viewer after upgrading to RStudio 1.1 and applying a
dark them, ostensibly because they're getting a cached 1.0 copy of the
styles. This change updates the CSS filenames (and others in the folder)
to ensure that 1.0 copies won't be used with 1.1.