update cached R Markdown package state after install; fixes #1801
After successfully installing the R Markdown package on demand, most of
the session info still indicates that the package doesn't exist. This
change updates the session info in place when a recent version of the
package is successfully installed so that commands like Knit with
Parameters can be used in the very first RStudio session.