Fix crash in core unit test on Mac
Not sure why this is only manifesting on qtwebengine branch on Mac; looks dubious on any platform. Not sure if it's a real-world crash case.

Anyway, the specific crash happened when calling `isPipeOperator(cursor.previousSignificantToken())` when `cursor` was already on the first token.

Under the covers this causes previousSignificantToken to return dummyToken, which is of type ERR, and cannot safely have its contents examined since it doesn't refer to anything. Trying to do so was putting code in unsafe state (including using -1 as a index into an unchecked array and possibly trying to use null iterator).

Fixed this specific issue by having isPipeOperator immediately bail on a token of type ERR.

While I was there, set default values on RToken object construction for members that were previously being left undefined in some cases.

It is possible there are other unsafe uses of previousSignificantToken's results. I skimmed and many of them already check the type of the resulting token, but I didn't do an exhaustive examination.