Reconnect terminal if websocket was closed
Fixes #1844

If the websocket was closed (seeing this happen in container scenarios), mark the terminal as disconnected so the reconnection logic kicks in. For example, terminal should reconnect if the user types in the disconnected terminal.

Prior to this fix, the terminal would remain inactive and ignore input if the underlying websocket had closed.