Show menu on satellite windows, not just main window (Mac specific)
Qt does extra Mac stuff to determine the menu shown for focused window.

Creating via QMenuBar() causes the placeholder menu to become the default for all windows; creating via QMenuBar(this) causes it to be associated with the current window, and it gets cleaned up later via menuBarCompleted signal and replaced with real menu bar (for all windows).