Avoid crash with native target=_blank links
When RStudio UI contains an anchor with target=_blank, it launches it in the system browser. In QtWebEngine the order of messages has changed, and a new QWebEngine page is first created and then sent an acceptNavigationRequest message, where we intercept the _blank and send to browser. This creates a new RStudio window, which was being displayed, then crashing upon its immediate destruction. Now I make that window hidden, and let it trigger the external load.