Add StringUtil.equals method for null-safe string comparison
Also, discovered that the core/client unit tests (such as for StringUtil) were not being run. Modified build.xml to include them.

Per a recent conversation, we decided to steer in the direction of using Java-compatible string comparision conventions, even though GWT output makes str1 == str2 an ok way to compare contents of strings (unlike Java).

For cases where null is a factor, added `StringUtil.equals(str1, str2)`. This defers to the newer Java `Objects.compare(str1, str2)`` method, but has the added benefit of compile-type type-safety (and gives a place to change our approach, should we need to).

For comparisions where we know the strings are not-null, then continued use of `str1.compare(str2)`` is the way to go, avoiding the slight performance penalty of the null checks inside Objects.compare.

Changed the Objects.equals I had previously added when IntelliJ complained about it. Did not do a sweep looking for `==` usages.