Stop Ctrl+L from sending focus to browser address when in Terminal
On Chrome (Windows and Linux, not Mac) hitting Ctrl+L in the RStudio terminal clears the terminal (as it should), but then puts focus in the browser address bar. This is quite painful, especially if you are a fast typist, hit Ctrl+L in terminal, then type a command and hit enter; now the browser is going somewhere you don't want it to.

Resolves #1636.