Fix issue with "less" and "more" handing the terminal
There was a bug in xterm.js, whereby if the terminal pane had fewer than 25 rows, and the first full-screen program run tried to do scrolling via linefeeds, it would cause JavaScript exceptions and break rendering. Only way to recover was to "q" less/more then refresh browser to get the main terminal buffer displayed again.

Made a fix to xterm.js 2.9.2 and applied to our own fork of xterm.js.

If and when we got to xterm.js 3.0, they have already fixed this bug, but aren't making more changes to 2.9.x.

Fixes #1875