remove unnecessary call to macOS 10.12 plus API
Early in the move back to Qt on the Mac, was getting View menu items
related to native Mac tab support; this call was hiding those. The
property, however, was added in 10.12, so prevented building on 10.11.

Looking now, it seems the newer version of Qt we're using has fixed this
for us, so removing the call altogether.

Fixes #2693