Don't show publish button in swagger pane or window
The publish button behavior inherited from Shiny implementation depends
on Shiny changing the working directory to the one containing the
sources; Plumber doesn't do this, and the result is that the publish
button will show everything in the current-working directory, which
may not be the right location. So, for Plumber, just have Publish
available in the editor toolbar.