Add config option to send package build output to package folder
Normally the Check Package, Build Source Package, and Build Binary Package commands write to the parent folder of the package itself. So for /home/me/mypackage, files are written to /home/me.

This change supports writing these files inside the package folder itself, for a case where users don't have write access to the folder containing the package project.

This can be enabled in rsession.conf by adding:

package-output-to-package-folder=1

This only works if package operations are being done via devtools package; if they are being done directly via R CMD check/install/build, this setting will not change the output location.

In addition to changing the output location, the .gitignore and .Rbuildignore files have additional exlusions added, covering the files generated by these operations.

These exclusions will be added only when this setting is enabled, and can happen when first creating a project or when opening an existing project.