Allow publishing rendered content from the Source pane
In RStudio 1.0 you could publish rendered content from the Source pane. The IDE would auto-run rendering before showing the publishing wizard (if current output was missing or out-of-date) to ensure it was available if user went down a publish-wizard path that needed it. However, publishing source to Connect does not need it, and often rendering on the client is not going to work at all for Connect users, so it was confusing and time-consuming.

RStudio 1.1 removed the ability to publish rendered content from the Source pane. It would only publish source from editor (i.e. publish source to Connect). To publish rendered content, had to use the Publish button on the rendered preview in viewer/window. Though this is logical, many users were used to using the editor's publish button and thought the ability to publish to RPubs or publish output-only to Connect had been removed.

This change puts back the ability to publish content from the Source pane. However, it does not do auto-render as in 1.0 to avoid the problems mentioned. Per conversation with @jmcphers, chose to show an error message if user goes down a publish path requiring rendered content, and it was missing, telling them to go do the render and try again. Injecting auto-render into the appropriate place in publishing wizard was deemed too messy, at least for 1.2.

Also detect when an attempt is made to publish unsupported rendered output to RPubs (e.g. PDF, MS-Word), instead of letting the deployment package shown an obscure error in the console.

Did not do the work to support publishing of other static-content formats such as PDF, Word, PPT, directly from the IDE. Broke that out into a separate issue to work on next.

Fixes #2816