switch Windows dockerfile back to static download of Qt
Installing Qt via scripted online installer in Windows Docker container being built on Server Core was failing, and could find no way to usefully debug it. So for Windows, use an archive built by installing on a clean machine using the Windows powershell bootstrapper script, and and uploaded to buildtools.

This adds a step to updating Qt version, updated the wiki to reflect this.