can't safely get address of empty vector
Ran into this doing Pro Server work on an Ubuntu Bionic (18) VM, with a Debug build of the product.

Server would assert and terminate on startup. This is same issue that httpuv project was hitting here:

https://github.com/rstudio/httpuv/issues/133

Debug build uses _GLIBCXX_ASSERTIONS and which asserts on usage of this pattern:

&v[0], where "v" is a vector

Take address of first element of a potentially empty vector is undefined behavior in C++.

Also can't just use vector::data(), because the return value of data() may or may not be nullptr for an empty vector, per C++ docs. (D'oh)

So added `safe_vec_addr` as used by httuv. I only fixed the one case that was preventing me from starting the server with a Debug build.

I will open an issue to do a comprehensive survey of the codebase to find other potential problems of this sort and/or fix them as I hit them.