Fixes inline latex math rendering
Inline math in this Shiny app fails, but only when tested on RStudio Server.
The correct behavior is for the math to be formatted by MathJax, but on
RStudio Server you see latex markup surrounded by $. The issue seems to be
that the copy of Sundown in rstudio is overriding the copy of Sundown in
the Markdown package, and only the latter contains the logic required to
process inline latex math. @jjallaire said that the intention was to keep
the two in sync, so this commit copies the relevant files over from markdown.
https://github.com/rstudio/shiny-examples/tree/master/020-knit-html